+++ GoGoCar/app/src/main/java/com/aristy/gogocar/FragmentApp.java
@@ -5,11 +5,13 @@ import android.os.Bundle;
 import androidx.fragment.app.Fragment;
 
 import android.os.Handler;
+import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
+import android.webkit.WebViewClient;
 
 import java.sql.Connection;
 
@@ -35,14 +37,37 @@ public class FragmentApp extends Fragment {
         // Inflate the layout for this fragment
         View view = inflater.inflate(R.layout.fragment_app, container, false);
 
+        // find items
         WebView web = view.findViewById(R.id.web_view);
         web.loadUrl("file:///android_asset/pages/home.html");
 
+        // Enable javascript
         WebSettings webSettings = web.getSettings();
         webSettings.setJavaScriptEnabled(true);
+
+        // Result state page
+        web.setWebViewClient(new Callback());
 
         web.addJavascriptInterface(new WebInterface(getActivity(), getContext(), web, SQLConnection, userPreferences, fragmentHandler), "Android");
 
         return view;
+    }
+
+    //open in app
+    public static class Callback extends WebViewClient {
+        @Override
+        public boolean shouldOverrideKeyEvent(WebView view, KeyEvent event) {
+            return false;
+        }
+
+        public void onPageFinished(WebView view, String url){
+            //Here you want to use .loadUrl again
+            //on the webView object and pass in
+            //"javascript:<your javaScript function"
+            //Set<BluetoothDevice> bluetoothDevice = getBluetoothPairedDevices();
+            //error here when bt is not activated
+            //populateSpinner(bluetoothDevice);
+        }
+
     }
 }
